requirements.txt
Django==1.6.5
Pillow==2.5.3
South==1.0
WebOb==1.4
beautifulsoup4==4.3.2
boto==2.32.1
dj-database-url==0.3.0
dj-static==0.0.6
django-allauth==0.18.0
django-braces==1.4.0
django-crispy-forms==1.4.0
django-mailgun==0.2.2
django-storages==1.1.8
django-tastypie==0.12.0
django-toolbelt==0.0.1
gunicorn==19.0.0
oauthlib==0.6.3
psycopg2==2.5.3
python-dateutil==2.2
python-mimeparse==0.1.4
python-openid==2.2.5
requests==2.4.1
requests-oauthlib==0.4.1
six==1.8.0
static3==0.5.1
waitress==0.8.9
wsgiref==0.1.2
